requirements.txt
dxcam==0.0.5
pyautogui==0.9.54
pygetwindow==0.0.9
pynput==1.7.6
pillow>=10.0.0
keyboard==0.13.5
numpy>=1.26.4
torch==2.3.0 --extra-index-url https://download.pytorch.org/whl/cpu